cssom-1_dev/xhtml1/cssstylerule.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSSOM CSSRule CSSStyleRule interface</title>
    <link href="mailto:lew.letitia@gmail.com" rel="author" title="Letitia Lew" />
    <link href="http://www.w3.org/TR/cssom-1/#css-rules" rel="help" />
    <link href="http://www.w3.org/TR/cssom-1/#the-cssrule-interface" rel="help" />
    <link href="http://www.w3.org/TR/cssom-1/#the-cssstylerule-interface" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="All properties for this CSSStyleRule instance of CSSRule are initialized correctly" name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <style type="text/css" id="styleElement">
        div { margin: 10px; padding: 0px; }
    </style>
</head>
<body>
    <div id="log"></div>

    <script type="text/javascript">
        var rule;
        setup(function() {
            var styleSheet = document.getElementById("styleElement").sheet;
            var ruleList = styleSheet.cssRules;
            rule = ruleList[0];
        });

        test(function() {
            assert_true(rule instanceof CSSRule);
            assert_true(rule instanceof CSSStyleRule);
        }, "CSSRule and CSSStyleRule types");

        test(function() {
            assert_equals(rule.STYLE_RULE, 1);
            assert_equals(rule.IMPORT_RULE, 3);
            assert_equals(rule.MEDIA_RULE, 4);
            assert_equals(rule.FONT_FACE_RULE, 5);
            assert_equals(rule.PAGE_RULE, 6);
            assert_equals(rule.NAMESPACE_RULE, 10);
            assert_idl_attribute(rule, "type");
            assert_equals(typeof rule.type, "number");
        }, "Type of CSSRule#type and constant values");

        test(function() {
            assert_true(rule instanceof CSSRule);
            assert_idl_attribute(rule, "cssText");
            assert_idl_attribute(rule, "parentRule");
            assert_idl_attribute(rule, "parentStyleSheet");
        }, "Existence of CSSRule attributes");

        test(function() {
            assert_readonly(rule, "type");
            assert_readonly(rule, "parentRule");
            assert_readonly(rule, "parentStyleSheet");
        }, "Writability of CSSRule attributes");

        test(function() {
            assert_equals(rule.type, rule.STYLE_RULE);
            assert_equals(typeof rule.cssText, "string");
            assert_equals(rule.cssText, "div { margin: 10px; padding: 0px; }");
            assert_equals(rule.parentRule, null);
            assert_true(rule.parentStyleSheet instanceof CSSStyleSheet);
        }, "Values of CSSRule attributes");

        test(function() {
            assert_idl_attribute(rule, "selectorText");
            assert_equals(typeof rule.selectorText, "string");
            assert_idl_attribute(rule, "style");
            assert_readonly(rule, "style");
        }, "Existence, writability and type of CSSStyleRule attributes");

        test(function() {
            assert_equals(rule.selectorText, "div");
            assert_true(rule.style instanceof CSSStyleDeclaration);
        }, "Values of CSSStyleRule attributes");
    </script>


</body></html>
